backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
policyengine-us
pydantic-settings
